Add nth, nunique, groups tests
