+++ python/cudf/cudf/tests/test_groupby.py
@@ -692,109 +692,6 @@ def test_groupby_apply_return_reindexed_series(as_index):
     assert_groupby_results_equal(expect, got)
 
 
-@pytest.mark.parametrize("agg", [lambda x: x.nunique(), "nunique"])
-@pytest.mark.parametrize("by", ["a", ["a", "b"], ["a", "c"]])
-def test_groupby_nunique(agg, by):
-    pdf = pd.DataFrame(
-        {"a": [1, 1, 1, 2, 3], "b": [1, 2, 2, 2, 1], "c": [1, 2, None, 4, 5]}
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    expect = pdf.groupby(by).nunique()
-    got = gdf.groupby(by).nunique()
-
-    assert_groupby_results_equal(expect, got, check_dtype=False)
-
-
-@pytest.mark.parametrize("dropna", [True, False])
-def test_nunique_dropna(dropna):
-    gdf = cudf.DataFrame(
-        {
-            "a": [1, 1, 2],
-            "b": [4, None, 5],
-            "c": [None, None, 7],
-            "d": [1, 1, 3],
-        }
-    )
-    pdf = gdf.to_pandas()
-
-    result = gdf.groupby("a")["b"].nunique(dropna=dropna)
-    expected = pdf.groupby("a")["b"].nunique(dropna=dropna)
-    assert_groupby_results_equal(result, expected, check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "n",
-    [0, 1, 2, 10],
-)
-@pytest.mark.parametrize("by", ["a", ["a", "b"], ["a", "c"]])
-def test_groupby_nth(n, by):
-    pdf = pd.DataFrame(
-        {
-            "a": [1, 1, 1, 2, 3],
-            "b": [1, 2, 2, 2, 1],
-            "c": [1, 2, None, 4, 5],
-            "d": ["a", "b", "c", "d", "e"],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    expect = pdf.groupby(by).nth(n)
-    got = gdf.groupby(by).nth(n)
-
-    assert_groupby_results_equal(expect, got, check_dtype=False)
-
-
-def test_raise_data_error():
-    pdf = pd.DataFrame({"a": [1, 2, 3, 4], "b": ["a", "b", "c", "d"]})
-    gdf = cudf.from_pandas(pdf)
-
-    assert_exceptions_equal(
-        pdf.groupby("a").mean,
-        gdf.groupby("a").mean,
-    )
-
-
-def test_multi_agg():
-    gdf = cudf.DataFrame(
-        {"a": [1, 1, 2, 2], "b": [1, 2, 3, 4], "c": ["a", "b", "c", "d"]}
-    )
-    pdf = gdf.to_pandas()
-    assert_groupby_results_equal(
-        pdf.groupby("a").agg({"b": ["count", "mean"], "c": ["count"]}),
-        gdf.groupby("a").agg({"b": ["count", "mean"], "c": ["count"]}),
-    )
-
-
-@pytest.mark.parametrize(
-    "agg",
-    (
-        [
-            *itertools.combinations(["count", "max", "min", "nunique"], 2),
-            {"b": "min", "c": "mean"},
-            {"b": "max", "c": "mean"},
-            {"b": "count", "c": "mean"},
-            {"b": "nunique", "c": "mean"},
-        ]
-    ),
-)
-def test_groupby_agg_combinations(agg):
-    pdf = pd.DataFrame(
-        {
-            "a": [1, 1, 2, 2, 3],
-            "b": ["a", "a", "b", "c", "d"],
-            "c": [1, 2, 3, 4, 5],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    assert_groupby_results_equal(
-        pdf.groupby("a").agg(agg),
-        gdf.groupby("a").agg(agg),
-        check_dtype=False,
-    )
-
-
 @pytest.mark.skipif(
     PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
     reason="Include groups missing on old versions of pandas",
@@ -816,114 +713,6 @@ def test_groupby_apply_noempty_group():
         .reset_index(drop=True)
     )
     assert_groupby_results_equal(expect, got)
-
-
-def test_reset_index_after_empty_groupby():
-    # GH #5475
-    pdf = pd.DataFrame({"a": [1, 2, 3]})
-    gdf = cudf.from_pandas(pdf)
-
-    assert_groupby_results_equal(
-        pdf.groupby("a").sum().reset_index(),
-        gdf.groupby("a").sum().reset_index(),
-        as_index=False,
-        by="a",
-    )
-
-
-def test_groupby_attribute_error():
-    err_msg = "Test error message"
-
-    class TestGroupBy(cudf.core.groupby.GroupBy):
-        @property
-        def _groupby(self):
-            raise AttributeError(err_msg)
-
-    a = cudf.DataFrame({"a": [1, 2], "b": [2, 3]})
-    gb = TestGroupBy(a, a["a"])
-
-    with pytest.raises(AttributeError, match=err_msg):
-        gb.sum()
-
-
-@pytest.mark.parametrize(
-    "by",
-    [
-        "a",
-        "b",
-        ["a"],
-        ["b"],
-        ["a", "b"],
-        ["b", "a"],
-        np.array([0, 0, 0, 1, 1, 1, 2]),
-    ],
-)
-def test_groupby_groups(by):
-    pdf = pd.DataFrame(
-        {"a": [1, 2, 1, 2, 1, 2, 3], "b": [1, 2, 3, 4, 5, 6, 7]}
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    pdg = pdf.groupby(by)
-    gdg = gdf.groupby(by)
-
-    for key in pdg.groups:
-        assert key in gdg.groups
-        assert_eq(pdg.groups[key], gdg.groups[key])
-
-
-@pytest.mark.parametrize(
-    "by",
-    [
-        "a",
-        "b",
-        ["a"],
-        ["b"],
-        ["a", "b"],
-        ["b", "a"],
-        ["a", "c"],
-        ["a", "b", "c"],
-    ],
-)
-def test_groupby_groups_multi(by):
-    pdf = pd.DataFrame(
-        {
-            "a": [1, 2, 1, 2, 1, 2, 3],
-            "b": ["a", "b", "a", "b", "b", "c", "c"],
-            "c": [1, 2, 3, 4, 5, 6, 7],
-        }
-    )
-    gdf = cudf.from_pandas(pdf)
-
-    pdg = pdf.groupby(by)
-    gdg = gdf.groupby(by)
-
-    for key in pdg.groups:
-        assert key in gdg.groups
-        assert_eq(pdg.groups[key], gdg.groups[key])
-
-
-def test_groupby_nunique_series():
-    pdf = pd.DataFrame({"a": [1, 1, 1, 2, 2, 2], "b": [1, 2, 3, 1, 1, 2]})
-    gdf = cudf.from_pandas(pdf)
-
-    assert_groupby_results_equal(
-        pdf.groupby("a")["b"].nunique(),
-        gdf.groupby("a")["b"].nunique(),
-        check_dtype=False,
-    )
-
-
-@pytest.mark.parametrize("list_agg", [list, "collect"])
-def test_groupby_list_simple(list_agg):
-    pdf = pd.DataFrame({"a": [1, 1, 1, 2, 2, 2], "b": [1, 2, None, 4, 5, 6]})
-    gdf = cudf.from_pandas(pdf)
-
-    assert_groupby_results_equal(
-        pdf.groupby("a").agg({"b": list}),
-        gdf.groupby("a").agg({"b": list_agg}),
-        check_dtype=False,
-    )
 
 
 @pytest.mark.parametrize("list_agg", [list, "collect"])
